+++ src/components/sign-in-page/sign-in-page.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import {Link} from "react-router-dom";
+import PageFooter from "../page-footer/page-footer.jsx";
 
 const SignInPage = () => {
   return (
@@ -36,19 +37,7 @@ const SignInPage = () => {
         </form>
       </div>
 
-      <footer className="page-footer">
-        <div className="logo">
-          <a href="#" className="logo__link logo__link--light">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </a>
-        </div>
-
-        <div className="copyright">
-          <p>© 2019 What to watch Ltd.</p>
-        </div>
-      </footer>
+      <PageFooter/>
     </div>
   );
 };
